+++ packages/rabbithole/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @rabbitholegg/questdk-plugin-rabbithole
 
+## 1.0.0-alpha.42
+
+### Minor Changes
+
+- [#385](https://github.com/rabbitholegg/questdk-plugins/pull/385) [`72a86095`](https://github.com/rabbitholegg/questdk-plugins/commit/72a860954036856b073a2043dfc8614c200585d6) Thanks [@sammccord](https://github.com/sammccord)! - use vite for build tooling, reduce plugin size, improve build speeds
+
+### Patch Changes
+
+- Updated dependencies [[`1fa4bb52`](https://github.com/rabbitholegg/questdk-plugins/commit/1fa4bb5262cef6a1f434850b097b9aacc0d1e8ba), [`19d8db5c`](https://github.com/rabbitholegg/questdk-plugins/commit/19d8db5c5cafb4409dd469d22bdfc709d6d108aa), [`72a86095`](https://github.com/rabbitholegg/questdk-plugins/commit/72a860954036856b073a2043dfc8614c200585d6)]:
+  - @rabbitholegg/questdk-plugin-utils@1.0.0-alpha.32
+  - @rabbitholegg/questdk@2.0.0-alpha.52
+
+## 1.0.0-alpha.41
+
+### Patch Changes
+
+- Updated dependencies [[`0d3c1441`](https://github.com/rabbitholegg/questdk-plugins/commit/0d3c1441f5eff0eaab84b414e26d6dee3ae1b590)]:
+  - @rabbitholegg/questdk-plugin-utils@1.0.0-alpha.31
+  - @rabbitholegg/questdk@2.0.0-alpha.51
+
 ## 1.0.0-alpha.40
 
 ### Patch Changes